[UP]Token CSRF

+++ src/Controller/Admin/Article/DeleteArticleController.php
@@ -4,7 +4,6 @@
 namespace App\Controller\Admin\Article;
 
 use App\Entity\Article;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -25,9 +24,11 @@ class DeleteArticleController extends AbstractController
      */
     public function delete(Request $request, Article $article): Response
     {
-//        if (!$this->isCsrfTokenValid('delete', $request->request->get('token'))) {
-//            return $this->redirectToRoute('app_home');
-//        }
+        $submittedToken = $request->get('token');
+
+        if (!$this->isCsrfTokenValid('admin_post_delete', $submittedToken)) {
+            return $this->redirectToRoute('app_home');
+        }
 
         $em = $this->getDoctrine()->getManager();
         $em->remove($article);
